% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{pay_select_dane_surowe}
\alias{pay_select_dane_surowe}
\title{Surowe dane ocen systemów płatności (PaySelectR)}
\format{
Ramka danych (data.frame) z 90 wierszami i 10 kolumnami:
\describe{
  \item{EkspertID}{Unikalny identyfikator eksperta (1-15)}
  \item{Alternatywa}{Nazwa systemu płatności (np. PayU, Stripe)}
  \item{prowizje}{Wysokość prowizji w procentach}
  \item{integracja_latwosc}{Ocena łatwości integracji (1-9)}
  \item{szybkosc_platnosci}{Ocena szybkości procesowania transakcji (1-9)}
  \item{bezpieczenstwo}{Ocena poziomu bezpieczeństwa (1-9)}
  \item{obsluga_klienta}{Ocena jakości wsparcia technicznego (1-9)), zawiera 99 jako błąd}
  \item{zasieg_miedzynarodowy}{Dostępność na rynkach zagranicznych (1-9)}
  \item{akceptacja_transakcji}{Procentowy wskaźnik udanych transakcji}
  \item{implementacja_latwosc}{Ocena łatwości wdrożenia przez deweloperów (1-9), zawiera braki danych NA}
}
}
\usage{
data(pay_select_dane_surowe)
}
\description{
Zbiór danych zawierający symulowane oceny 15 ekspertów dla 6 systemów płatności
w oparciu o 8 kluczowych kryteriów e-commerce.
}
