% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_vikor_plot.R
\name{plot.pay_fuzzy_vikor_res}
\alias{plot.pay_fuzzy_vikor_res}
\title{Plot Fuzzy VIKOR Results}
\usage{
\method{plot}{pay_fuzzy_vikor_res}(x, ...)
}
\arguments{
\item{x}{An object of class `pay_fuzzy_vikor_res` returned by the `pay_fuzzy_vikor()` function.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Visualizes the VIKOR results using a bubble chart.
The X-axis represents UtilityScore, the Y-axis represents RiskScore.
Bubble size is inversely proportional to the FinalScore - smaller (better rank) = larger bubble.
}
